move auth router to /routers
