+++ backend/src/routers/auth-http-router.ts
@@ -1,7 +1,7 @@
 import { userDataResponseSchema } from 'digital-fuesim-manv-shared';
 import { Router } from 'express';
+import type { AuthService } from 'auth/auth-service.js';
 import { toFrontend } from '../utils/frontend-http-helper.js';
-import type { AuthService } from './auth-service.js';
 
 export const createAuthRouter = (authService: AuthService) => {
     const router = Router();
